consumer_elastic/requirements.txt
kafka-python==2.0.2
elasticsearch==8.0.0
requests==2.28.1
